catching up with upstream

+++ bin/archive.sh
@@ -2,19 +2,33 @@
 #--
 # archive.sh
 #--
-zipfile=aws-extend-switch-roles.zip
+copydest=$1
 
 cd dist/chrome;
-\rm $zipfile
+version=`cat manifest.json | jq -r '.version' | sed 's/\./-/g'`
+zipfile="aesr-chrome-$version.zip"
+if [ -e $zipfile ]; then
+  \rm $zipfile
+fi
 zip -r $zipfile \
   manifest.json *.html icons/ js/ 
 echo "archived: chrome/$zipfile"
+if [ "$copydest" != "" ]; then
+  \cp $zipfile $copydest
+fi
 
 echo "----"
 
 cd ../firefox;
-\rm $zipfile
+version=`cat manifest.json | jq -r '.version' | sed 's/\./-/g'`
+zipfile="aesr-firefox-$version.zip"
+if [ -e $zipfile ]; then
+  \rm $zipfile
+fi
 zip -r $zipfile \
   manifest.json *.html icons/ js/ 
 
 echo "archived: firefox/$zipfile"
+if [ "$copydest" != "" ]; then
+  \cp $zipfile $copydest
+fi
